+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 ChangeLog
 ---------
 
+### 1.13.5
+
+- **New(auto-refresh extension):** Rewrote auto-refresh extension to ES6.
+- **Update(js):** Fixed showFullscreen cannot work bug.
+- **Update(js):** Redefined customSearch option.
+- **Update(js):** Fixed show footer cannot work bug.
+- **Update(js):** Updated the parameter of `footerStyle`.
+- **Update(js):** Added classes supported for `footerStyle`.
+- **Update(js):** Fixed IE11 transform bug.
+- **Update(js):** Removed beginning and end whitespace from td.
+- **Update(export extension):** Fixed export selected bug.
+
 ### 1.13.4
 
 - **New(sticky-header extension):** Rewrote sticky-header extension to ES6.